requirements.txt
# web framework
fastapi==0.115.8
uvicorn[standard]==0.34.0

# database
sqlalchemy==2.0.36
alembic==1.14.0

# configuration
python-dotenv==1.0.1

# data validation
pydantic==2.10.6
pydantic-settings==2.1.0

# cache
redis==5.2.1

# http client
aiohttp==3.9.1

# testing
backoff==2.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# http client
httpx==0.25.2

# http client
python-multipart==0.0.6

# template engine
jinja2==3.1.2

# hashids
hashids==1.3.1

# database
asyncpg==0.29.0
